+++ frontend/src/App.tsx
@@ -6,6 +6,7 @@ import HomePage from "./containers/HomePage.tsx";
 import Header from "./components/Common/Header.tsx";
 import Footer from "./components/Common/Footer.tsx";
 import ArticlePage from "./containers/Article/ArticlePage.tsx";
+import OverviewPage from "./containers/Article/OverviewPage.tsx";
 
 const LoginPage = lazy(() => import("./containers/User/LoginPage.tsx"));
 const ArticleMainPage = lazy(
@@ -46,7 +47,25 @@ function App() {
           />
 
           <Route
+            path="/article/draft"
+            element={
+              <Suspense fallback={<div>Loading...</div>}>
+                <ArticlePage />
+              </Suspense>
+            }
+          />
+
+          <Route
             path="/article/:id"
+            element={
+              <Suspense fallback={<div>Loading...</div>}>
+                <OverviewPage />
+              </Suspense>
+            }
+          />
+
+          <Route
+            path="/article/:id/edit"
             element={
               <Suspense fallback={<div>Loading...</div>}>
                 <ArticlePage />
