+++ frontend/src/components/Common/MDX/allPlugins.ts
@@ -1,0 +1,54 @@
+import {
+  toolbarPlugin,
+  codeBlockPlugin,
+  sandpackPlugin,
+  codeMirrorPlugin,
+  headingsPlugin,
+  listsPlugin,
+  quotePlugin,
+  thematicBreakPlugin,
+  markdownShortcutPlugin,
+  linkPlugin,
+  linkDialogPlugin,
+  imagePlugin,
+  tablePlugin,
+  frontmatterPlugin,
+  directivesPlugin,
+  AdmonitionDirectiveDescriptor,
+  diffSourcePlugin,
+} from "@mdxeditor/editor";
+import { reactSandpackConfig } from "./reactSandpackConfig";
+
+export const allPlugins = (diffMarkdown: string) => [
+  toolbarPlugin({
+    toolbarContents: () => {
+      return "";
+    },
+  }),
+  listsPlugin(),
+  quotePlugin(),
+  headingsPlugin(),
+  linkPlugin(),
+  linkDialogPlugin(),
+  // eslint-disable-next-line @typescript-eslint/require-await
+  imagePlugin({}),
+  tablePlugin(),
+  thematicBreakPlugin(),
+  frontmatterPlugin(),
+  codeBlockPlugin({ defaultCodeBlockLanguage: "txt" }),
+  sandpackPlugin({ sandpackConfig: reactSandpackConfig }),
+  codeMirrorPlugin({
+    codeBlockLanguages: {
+      js: "JavaScript",
+      css: "CSS",
+      txt: "text",
+      tsx: "TypeScript",
+      cplusplus: "C++",
+      java: "Java",
+      c: "C",
+    },
+  }),
+  directivesPlugin({ directiveDescriptors: [AdmonitionDirectiveDescriptor] }),
+  diffSourcePlugin({ viewMode: "rich-text", diffMarkdown }),
+  markdownShortcutPlugin(),
+];
